CDSR-787 Making use of Multiple MRNs summary
