+++ conf/messages
@@ -973,7 +973,6 @@ check-your-answers.basis.h2=Basis for claim
 check-your-answers.basis.l0=This is the basis behind the claim
 check-your-answers.northern-ireland-claim.label=Were your goods imported into Northern Ireland?
 check-your-answers.commodities-details.label=This is the reason for the claim
-check-your-answers.commodities-details.scheduled.label=This is the reason for the claim
 check-your-answers.claim-calculation.label=Claim calculation
 check-your-answers.claim-calculation.l0=Amount of UK Duty to reimburse claimed by you:
 check-your-answers.claim-calculation.l1=Amount of EU Duty to reimburse claimed by you:
@@ -998,7 +997,6 @@ check-your-answers.reference-number.scheduled.h2=Lead Movement Reference Number 
 check-your-answers.declaration-details.h2=Declaration details
 check-your-answers.northern-ireland-claim.h2=Were your goods imported into Northern Ireland?
 check-your-answers.commodity-details.h2=Reason for claim
-check-your-answers.commodity-details.scheduled.h2=Reason for claim
 check-your-answers.claim-calculation.h2=Total reimbursement claim
 check-your-answers.bank-details.h2=Bank details
 check-your-answers.attached-documents.h2=Supporting documents
