+++ test/uk/gov/hmrc/cdsreimbursementclaimfrontend/controllers/claims/CheckYourAnswersSummarySpec.scala
@@ -16,23 +16,15 @@
 
 package uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims
 
-import cats.implicits.{catsSyntaxApply, catsSyntaxTuple2Semigroupal}
+import org.jsoup.nodes
 import org.scalatest.OptionValues
 import play.api.i18n.{Lang, Messages, MessagesApi, MessagesImpl}
 import play.api.inject.bind
 import play.api.inject.guice.GuiceableModule
-import play.api.test.FakeRequest
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.cache.SessionCache
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims.CheckYourAnswersAndSubmitController.checkYourAnswersKey
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims.SelectBasisForClaimController.selectBasisForClaimKey
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims.SelectWhoIsMakingTheClaimController.whoIsMakingTheClaimKey
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.fileupload.SupportingEvidenceController.supportingEvidenceKey
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.{AuthSupport, ControllerSpec, JourneyBindable, SessionSupport}
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.DeclarantTypeAnswer.{items => declarantTypes}
+import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.{AuthSupport, ControllerSpec, SessionSupport}
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.JourneyStatus.FillingOutClaim
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.SelectNumberOfClaimsAnswer.{Individual, Scheduled}
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.utils.MoneyUtils.formatAmountOfMoneyWithPoundSign
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.generators.DraftClaimGen._
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.generators.Generators.sample
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.generators.IdGen._
@@ -42,12 +34,9 @@ import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.{DraftClaim, SelectNumbe
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.services.ClaimService
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.support.HtmlParseSupport
 
-import scala.collection.JavaConverters._
-
-class CheckYourAnswersSummarySpec
+abstract class CheckYourAnswersSummarySpec
     extends ControllerSpec
     with OptionValues
-    with HtmlParseSupport
     with SessionSupport
     with AuthSupport {
 
@@ -60,365 +49,48 @@ class CheckYourAnswersSummarySpec
       bind[ClaimService].toInstance(mockClaimService)
     )
 
-  private val controller: CheckYourAnswersAndSubmitController = instanceOf[CheckYourAnswersAndSubmitController]
+  protected val controller: CheckYourAnswersAndSubmitController =
+    instanceOf[CheckYourAnswersAndSubmitController]
 
   implicit lazy val messagesApi: MessagesApi = controller.messagesApi
   implicit lazy val messages: Messages       = MessagesImpl(Lang("en"), messagesApi)
 
-  "The CYA page" should {
-
-    "display answer summaries for the Single journey" in {
-      val (session, claim) = genData(Individual)
-
-      inSequence {
-        mockAuthWithNoRetrievals()
-        mockGetSession(session)
-      }
-
-      val result = controller.checkAllAnswers(JourneyBindable.Single)(FakeRequest())
-
-      checkPageIsDisplayed(
-        result,
-        messageFromMessageKey(s"$checkYourAnswersKey.title"),
-        doc => {
-          val headers = doc.select("#main-content > div > div > h2").content
-
-          val summaries = doc
-            .select("#main-content > div > div > dl > div")
-            .asScala
-            .flatMap { element =>
-              val label      = element.select("dt").text()
-              val value      = element.select("dd").not(".govuk-summary-list__actions")
-              val paragraphs = value.select("p")
-
-              if (paragraphs.isEmpty)
-                Seq((label, value.text()))
-              else
-                paragraphs.content
-                  .map(s => (label, s.replace("<br>", " ")))
-            }
-            .toList
-
-          headers should contain allElementsOf (Seq(
-            claim.basisOfClaimAnswer *> Some(s"$checkYourAnswersKey.basis.h2"),
-            claim.displayDeclaration *> Some(s"$checkYourAnswersKey.declaration-details.h2"),
-            claim.mrnContactAddressAnswer *> claim.mrnContactDetailsAnswer *> Some(
-              s"$checkYourAnswersKey.contact-details.h2"
-            )
-          ).flatMap(_.toList) ++ Seq(
-            s"$checkYourAnswersKey.claimant-type.h2",
-            s"$checkYourAnswersKey.commodity-details.h2",
-            s"$checkYourAnswersKey.attached-documents.h2",
-            s"$checkYourAnswersKey.reference-number.h2"
-          )).map(messages(_))
-
-          summaries should contain allElementsOf Seq(
-            (
-              messages(s"$checkYourAnswersKey.claimant-type.l0"),
-              messages(
-                s"$whoIsMakingTheClaimKey.importer${declarantTypes.indexOf(claim.declarantTypeAnswer.value)}"
-              )
-            ),
-            (
-              messages(s"$checkYourAnswersKey.commodities-details.label"),
-              claim.commoditiesDetailsAnswer.map(_.value).value
-            ),
-            (
-              messages(s"$checkYourAnswersKey.reference-number.label"),
-              claim.movementReferenceNumber.value.value
-            )
-          ) ++ claim.basisOfClaimAnswer.map { answer =>
-            (
-              messages(s"$checkYourAnswersKey.basis.l0"),
-              messages(s"$selectBasisForClaimKey.reason.d${answer.value}")
-            )
-          }.toList ++ claim.supportingEvidencesAnswer.value.map { uploadDocument =>
-            (
-              messages(s"$checkYourAnswersKey.attached-documents.label"),
-              s"${uploadDocument.fileName} ${uploadDocument.documentType.fold("")(documentType =>
-                messages(s"$supportingEvidenceKey.choose-document-type.document-type.d${documentType.index}")
-              )}"
-            )
-          }.toList ++ claim.displayDeclaration.toList
-            .flatMap { declaration =>
-              Seq(
-                Some(
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.import-date-label"),
-                    declaration.displayResponseDetail.acceptanceDate
-                  )
-                ),
-                Some(
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.paid-charges-label"),
-                    formatAmountOfMoneyWithPoundSign(declaration.totalPaidCharges)
-                  )
-                ),
-                declaration.consigneeName.map { name =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.importer-name-label"),
-                    name
-                  )
-                },
-                declaration.consigneeEmail.map { email =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.importer-email-label"),
-                    email
-                  )
-                },
-                declaration.consigneeAddress.map { address =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.importer-address-label"),
-                    address.replace("<br />", " ")
-                  )
-                },
-                Some(
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.declarant-name-label"),
-                    declaration.declarantName
-                  )
-                ),
-                declaration.declarantContactAddress.map { address =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.declarant-address-label"),
-                    address.replace("<br />", " ")
-                  )
-                }
-              )
-            }
-            .flatMap(_.toList) ++ (claim.mrnContactDetailsAnswer, claim.mrnContactAddressAnswer)
-            .mapN { (details, address) =>
-              Seq(
-                Some((messages(s"claimant-details.contact.details"), details.fullName)),
-                details.phoneNumber.map { phoneNumber =>
-                  (messages(s"claimant-details.contact.details"), phoneNumber.value)
-                },
-                Some((messages(s"claimant-details.contact.details"), details.emailAddress.value)),
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    address.line1
-                  )
-                ),
-                address.line2.map { line2 =>
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    line2
-                  )
-                },
-                address.line3.map { line3 =>
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    line3
-                  )
-                },
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    address.line4
-                  )
-                ),
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    address.postcode
-                  )
-                ),
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    messages(address.country.messageKey)
-                  )
-                )
-              ).flatMap(_.toList)
-            }
-            .getOrElse(Seq.empty)
-        }
-      )
-    }
-
-    "display answer summaries for the Scheduled journey" in {
-      val (session, claim) = genData(Scheduled)
-
-      inSequence {
-        mockAuthWithNoRetrievals()
-        mockGetSession(session)
-      }
-
-      val result = controller.checkAllAnswers(JourneyBindable.Scheduled)(FakeRequest())
-
-      checkPageIsDisplayed(
-        result,
-        messageFromMessageKey(s"$checkYourAnswersKey.title"),
-        doc => {
-          val headers = doc.select("#main-content > div > div > h2").content
-
-          val summaries = doc
-            .select("#main-content > div > div > dl > div")
-            .asScala
-            .flatMap { element =>
-              val label      = element.select("dt").text()
-              val value      = element.select("dd").not(".govuk-summary-list__actions")
-              val paragraphs = value.select("p")
-
-              if (paragraphs.isEmpty)
-                Seq((label, value.text()))
-              else
-                paragraphs.content
-                  .map(s => (label, s.replace("<br>", " ")))
-            }
-            .toList
-
-          headers should contain allElementsOf (Seq(
-            claim.basisOfClaimAnswer *> Some(s"$checkYourAnswersKey.basis.h2"),
-            claim.displayDeclaration *> Some(s"$checkYourAnswersKey.declaration-details.h2"),
-            claim.mrnContactAddressAnswer *> claim.mrnContactDetailsAnswer *> Some(
-              s"$checkYourAnswersKey.contact-details.h2"
-            )
-          ).flatMap(_.toList) ++ Seq(
-            s"$checkYourAnswersKey.claimant-type.h2",
-            s"$checkYourAnswersKey.commodity-details.scheduled.h2",
-            s"$checkYourAnswersKey.attached-documents.h2",
-            s"$checkYourAnswersKey.scheduled-document.h2",
-            s"$checkYourAnswersKey.reference-number.scheduled.h2"
-          )).map(messages(_))
-
-          summaries should contain allElementsOf Seq(
-            (
-              messages(s"$checkYourAnswersKey.reference-number.scheduled.label"),
-              claim.movementReferenceNumber.value.value
-            ),
-            (
-              messages(s"$checkYourAnswersKey.claimant-type.l0"),
-              messages(
-                s"$whoIsMakingTheClaimKey.importer${declarantTypes.indexOf(claim.declarantTypeAnswer.value)}"
-              )
-            ),
-            (
-              messages(s"$checkYourAnswersKey.commodities-details.scheduled.label"),
-              claim.commoditiesDetailsAnswer.map(_.value).value
-            ),
-            (
-              messages(s"$checkYourAnswersKey.scheduled-document.label"),
-              claim.scheduledDocumentAnswer.map(_.uploadDocument.fileName).value
-            )
-          ) ++ claim.basisOfClaimAnswer.map { answer =>
-            (
-              messages(s"$checkYourAnswersKey.basis.l0"),
-              messages(s"$selectBasisForClaimKey.reason.d${answer.value}")
-            )
-          }.toList ++ claim.supportingEvidencesAnswer.value.map { uploadDocument =>
-            (
-              messages(s"$checkYourAnswersKey.attached-documents.label"),
-              s"${uploadDocument.fileName} ${uploadDocument.documentType.fold("")(documentType =>
-                messages(s"$supportingEvidenceKey.choose-document-type.document-type.d${documentType.index}")
-              )}"
-            )
-          }.toList ++ claim.displayDeclaration.toList
-            .flatMap { declaration =>
-              Seq(
-                Some(
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.import-date-label"),
-                    declaration.displayResponseDetail.acceptanceDate
-                  )
-                ),
-                Some(
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.paid-charges-label"),
-                    formatAmountOfMoneyWithPoundSign(declaration.totalPaidCharges)
-                  )
-                ),
-                declaration.consigneeName.map { name =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.importer-name-label"),
-                    name
-                  )
-                },
-                declaration.consigneeEmail.map { email =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.importer-email-label"),
-                    email
-                  )
-                },
-                declaration.consigneeAddress.map { address =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.importer-address-label"),
-                    address.replace("<br />", " ")
-                  )
-                },
-                Some(
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.declarant-name-label"),
-                    declaration.declarantName
-                  )
-                ),
-                declaration.declarantContactAddress.map { address =>
-                  (
-                    messages(s"$checkYourAnswersKey.declaration-details.declarant-address-label"),
-                    address.replace("<br />", " ")
-                  )
-                }
-              )
-            }
-            .flatMap(_.toList) ++ (claim.mrnContactDetailsAnswer, claim.mrnContactAddressAnswer)
-            .mapN { (details, address) =>
-              Seq(
-                Some((messages(s"claimant-details.contact.details"), details.fullName)),
-                details.phoneNumber.map { phoneNumber =>
-                  (messages(s"claimant-details.contact.details"), phoneNumber.value)
-                },
-                Some((messages(s"claimant-details.contact.details"), details.emailAddress.value)),
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    address.line1
-                  )
-                ),
-                address.line2.map { line2 =>
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    line2
-                  )
-                },
-                address.line3.map { line3 =>
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    line3
-                  )
-                },
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    address.line4
-                  )
-                ),
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    address.postcode
-                  )
-                ),
-                Some(
-                  (
-                    messages(s"claimant-details.contact.address"),
-                    messages(address.country.messageKey)
-                  )
-                )
-              ).flatMap(_.toList)
-            }
-            .getOrElse(Seq.empty)
-        }
-      )
-    }
-  }
-
-  private def genData(selectNumberOfClaimsAnswer: SelectNumberOfClaimsAnswer): (SessionData, DraftClaim) = {
+  protected def genData(selectNumberOfClaimsAnswer: SelectNumberOfClaimsAnswer): (SessionData, DraftClaim) = {
     val ggCredId            = sample[GGCredId]
     val signedInUserDetails = sample[SignedInUserDetails]
     val claim               = sample(genValidDraftClaim(selectNumberOfClaimsAnswer))
     val fillingOutClaim     = FillingOutClaim(ggCredId, signedInUserDetails, claim)
     val session             = SessionData.empty.copy(journeyStatus = Some(fillingOutClaim))
     (session, claim)
+  }
+}
+
+object CheckYourAnswersSummarySpec extends HtmlParseSupport {
+  import scala.collection.JavaConverters._
+
+  implicit class DOMDocOps(val document: nodes.Document) extends AnyVal {
+
+    def extractHeaders(): Seq[String] =
+      document
+        .select("#main-content > div > div > h2")
+        .content
+
+    def extractSummaries(): Seq[(String, String)] =
+      document
+        .select("#main-content > div > div > dl > div")
+        .asScala
+        .flatMap { element =>
+          val label      = element.select("dt").text()
+          val value      = element.select("dd").not(".govuk-summary-list__actions")
+          val paragraphs = value.select("p")
+
+          if (paragraphs.isEmpty)
+            Seq((label, value.text()))
+          else
+            paragraphs.content.map { s =>
+              (label, s.replace("<br>", " "))
+            }
+        }
+        .toList
   }
 }
